feat : Creator 평가들 반환 기능

+++ src/main/java/org/whitebox/howlook/domain/evaluation/dto/CreatorReviewDTO.java
@@ -17,7 +17,7 @@ public class CreatorReviewDTO {
     private String mainPhotoPath;
 
     // 매긴 점수
-    private Double score;
+    private Long score;
 
     // 남긴 리뷰
     private String review;
